benchmarking/requirements.txt
bnlearn
pandas
numpy
fastparquet
pyarrow
scipy
pgmpy
pyyaml
